+++ set-env.ts
@@ -15,19 +15,20 @@
  *   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
  */
 
-const fs = require("fs");
+const fs = require('fs');
 // Configure Angular `environment.ts` file path
-const target = "environment.prod.ts";
+const target = 'environment.prod.ts';
 const targetPath = `./src/environments/${target}`;
 // Load node modules
-const colors = require("colors");
-require("dotenv").load();
+const colors = require('colors');
+require('dotenv').load();
 // `environment.ts` file structure
 const envConfigFile = `export const environment = {
-  production: true,
+   production: true,
    apiUrl: '${process.env.API_URL}',
    identityServerUrl: '${process.env.IDS_URL}',
-};
+   sentryDsnUrl: '${process.env.SENTRY_DSN_URL}'
+  };
 `;
 console.log(colors.magenta(`The file ${target} will be written with the following content: \n`));
 console.log(colors.grey(envConfigFile));