+++ src/app/interceptors/http.interceptor.ts
@@ -1,0 +1,42 @@
+/*
+ *  Digital Excellence Copyright (C) 2020 Brend Smits
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU Lesser General Public License as published
+ *   by the Free Software Foundation version 3 of the License.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty
+ *   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *   See the GNU Lesser General Public License for more details.
+ *
+ *   You can find a copy of the GNU Lesser General Public License
+ *   along with this program, in the LICENSE.md file in the root project directory.
+ *   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
+ */
+import { HttpRequest, HttpInterceptor, HttpHandler, HttpEvent, HttpErrorResponse } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { Observable, EMPTY } from 'rxjs';
+import { catchError, retry } from 'rxjs/operators';
+import * as Sentry from '@sentry/browser';
+import { environment } from 'src/environments/environment';
+
+@Injectable()
+export class HttpErrorInterceptor implements HttpInterceptor {
+
+    constructor() { }
+
+    intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+        return next.handle(request).pipe(
+            retry(1),
+            catchError((error: HttpErrorResponse) => {
+                if (environment.production) {
+                    // Log error to sentry.
+                    Sentry.captureException(new Error(` http error:${error.status} - ${error.message}`));
+                }
+                // Stop error from continuing.
+                return EMPTY;
+            })
+        );
+    }
+}